+++ src/main/java/net/telepathicgrunt/ultraamplified/extrabehavior/PlayerSleepBehavior.java
@@ -1,20 +1,18 @@
 package net.telepathicgrunt.ultraamplified.extrabehavior;
 
-import net.minecraft.client.Minecraft;
 import net.minecraft.world.GameRules;
 import net.minecraft.world.IWorld;
 import net.minecraft.world.World;
 import net.minecraft.world.dimension.Dimension;
 import net.minecraft.world.dimension.DimensionType;
 import net.minecraft.world.server.ServerWorld;
-import net.minecraftforge.event.TickEvent.ClientTickEvent;
 import net.minecraftforge.event.world.SleepFinishedTimeEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.telepathicgrunt.ultraamplified.UltraAmplified;
-import net.telepathicgrunt.ultraamplified.world.dimension.UADimensionRegistration;
 import net.telepathicgrunt.ultraamplified.world.dimension.TimeSyncNetworkPacket;
 import net.telepathicgrunt.ultraamplified.world.dimension.UADimension;
+import net.telepathicgrunt.ultraamplified.world.dimension.UADimensionRegistration;
 
 
 @Mod.EventBusSubscriber(modid = UltraAmplified.MODID, bus = Mod.EventBusSubscriber.Bus.MOD)
@@ -49,30 +47,6 @@ public class PlayerSleepBehavior
 					{
 						((ServerWorld)world).getServer().getWorld(DimensionType.OVERWORLD).dimension.resetRainAndThunder();
 					}
-				}
-			}
-		}
-
-		
-		static int timeIncrementMod = 0;
-		
-		@SubscribeEvent
-		/**
-		 * Increments the time used for the daytime on client side.
-		 * Thus the day cycle appears smoother on the client.
-		 */
-		public static void IncrementTimeClientSided(ClientTickEvent event)
-		{
-			@SuppressWarnings("resource")
-			World world = Minecraft.getInstance().world;
-			if(world != null && !Minecraft.getInstance().isGamePaused() && world.getGameRules().getBoolean(GameRules.DO_DAYLIGHT_CYCLE)) 
-			{
-				Dimension dimension = world.dimension;
-				if(dimension.getType() == UADimensionRegistration.ultraamplified()) 
-				{
-					timeIncrementMod++;
-					if(timeIncrementMod % 2 == 0)
-						((UADimension)dimension).setWorldTimeClientSided(((UADimension)dimension).getWorldTime()+1L);
 				}
 			}
 		}